requirements.txt
openai==1.51.0
httpx==0.27.2
pandas==2.3.2
pdfplumber==0.11.7
python-dotenv==1.1.1
streamlit==1.39.0
supabase==2.0.0
